Migrated the testing patches classes to their own file
